Refactor for rust-analyzer bug

+++ badminton-tool/src-tauri/src/models/player.rs
@@ -34,7 +34,7 @@ pub struct CreatePlayerRequest {
     pub email: String,
     pub gender: Gender,
     pub club_id: i64,
-    pub skill_level: i8,
+    pub skill_level: i32,
 }
 
 impl CreatePlayerRequest {
